api/requirements.txt
fastapi
uvicorn
pydantic
pycryptodome
pyrebase4
